Rename multisearch widget plugin
** Why are these changes being introduced:

* We are renaming all our local plugins to use a standard naming pattern
  and metadata approach.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/lm-13

** How does this address that need:

* This renames the plugin to Mitlib Multisearch Widget - so the folder
  name changes, as well as the base PHP script and .gitignore.

** Document any side effects to this change:

* The CodeClimate badges are removed from the readme, and a license
  statement in a block comment of the base PHP script has been
  removed.

** Future work that I've resisted the urge to do:

* There are some other needed changes to this plugin (removing the
  option for the EDS templates, and moving to a namespace like
  Mitlib/Multisearch instead of just Mitlib), but I've resisted the urge
  to do that here.
* Similarly, we don't have a great way of dealing with JS dependencies
  for these plugins - so we're just committing the responsive tabs
  jQuery plugin in a libs directory. An alternative to this would be to
  check if someone has ported this plugin to Packagist somewhere, or to
  look back at that gist that was shared with us to bridge Composer and
  Grunt. But not now.
