+++ web/app/plugins/mitlib-multisearch-widget/README.md
@@ -1,0 +1,3 @@
+# MITlib Multisearch Widget
+
+This Wordpress plugin provides a search widget that can be used to direct searches against multiple targets. It is useful for libraries whose patrons want to search various resources, such as specialized databases, a central library catalog, or web-scale discovery service.
