+++ web/app/plugins/mitlib-multisearch-widget/i/down-caret.png
