+++ web/app/plugins/mitlib-multisearch-widget/mitlib-multisearch-widget.js
