--- web/app/plugins/wp-multisearch-widget/README.md
@@ -1,6 +1,0 @@
-# Multisearch Widget
-
-[![Code Climate](https://codeclimate.com/github/MITLibraries/wp-multisearch-widget/badges/gpa.svg)](https://codeclimate.com/github/MITLibraries/wp-multisearch-widget)
-[![Issue Count](https://codeclimate.com/github/MITLibraries/wp-multisearch-widget/badges/issue_count.svg)](https://codeclimate.com/github/MITLibraries/wp-multisearch-widget)
-
-This Wordpress plugin provides a search widget that can be used to direct searches against multiple targets. It is useful for libraries whose patrons want to search various resources, such as specialized databases, a central library catalog, or web-scale discovery service.
