--- web/app/plugins/wp-multisearch-widget/wp-multisearch-widget.php
@@ -1,47 +1,0 @@
-<?php
-/**
- * Plugin Name: Multisearch Widget
- * Plugin URI: https://github.com/MITLibraries/wp-multisearch-widget
- * Description: This plugin provides a widget that provides searches against multiple targets.
- * Version: 1.6.0
- * Author: MIT Libraries
- * Author URI: https://github.com/MITLibraries
- * License: GPL2
- *
- * @package Multisearch Widget
- * @author MIT Libraries
- * @link https://github.com/MITLibraries/wp-multisearch-widget
- */
-
-/**
- * Multisearch Widget is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 2 of the License, or
- * any later version.
- *
- * Multisearch Widget is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with Multisearch Widget. If not, see {URI to Plugin License}.
- */
-
-namespace mitlib;
-
-// Don't call the file directly!
-if ( ! defined( 'ABSPATH' ) ) {
-	exit;
-}
-
-// Include the necessary classes.
-include_once( 'class-multisearch-widget.php' );
-
-/**
- * Registers base widget.
- */
-function register_multisearch_widget() {
-	register_widget( 'mitlib\Multisearch_Widget' );
-}
-add_action( 'widgets_init', 'mitlib\register_multisearch_widget' );
